Week 4 progress
As of now I have the code to send the contact information but I am
currently trying to connect to the server to make this happen. The
connection is the hardest part so far, and I am currently watching
videos on the wiki in an effort to complete this task.