Week 7 progress
So far I have completed the canvas where the user can create the
skeleton frame work for their desired look of their wenbsite. I am
trying to adjust the height and width of the div box to be resizable by
the user and I plan to add options for the user to change the color.