Week 8 progress
I was able to change the background to a pattern I liked, however when I
tried to change the canvas to be draggable, I was not able to do so. I
tried to use the draggable textbox and apply it to the canvas but that
did not work. My main focus has been to try and get that working because
I figured that, this would be easier than setting up the color changer
for the squares. I also tried to change the position into auto, and that
will automatically extend the canvas, but it will not keep it in place
and keep the square centered.