Week 10 progress
So far I have still been trying to get the color to change, but I have
not been able to accomplish that yet. I had a problem where the box
would be extended past the border, which I fixed. But now I need to get
the color to actually change and for the color selctions to show up as
buttons, or be highlighted.