Week 12 progress
So far I went back to the scheduler and actually got the calendar to
show up, just need to work on getting the alert working for the user to
have confirmation that their meeting has been schduled. I also have been
workin gon getting the color to change and I feel that I am very close,
but I just cant get it exact. My hopes for this coming week is to finsh
the color change early on in the week and get the alert to show up.
possibly add a print screen function if time permits.