Week 13 progress
As of now I have the color in the square being filled, but the only
thing I need to do now is have it so that multiple squares with
different boxes appear at the same time. The way it is set up now, if
you select one color and have 3 boxes open at once, all the boxes will
change to that color. I need to try to get it to remain different colors
for the desired amount of boxes needed