Week 14 progress
Updated the scheduler and I actually got the button to work correctly;
Cleaned up the layout a bit; Currently working on having everything look
clean and have them appear to be functioning.  Also want the color in
the builder to appear in different squares, not just one. Going to try
to do a seperate css with classes and an array